Move retry logic into its own file
The retry logic has enough complexity for its own module.
