+++ crates/uv-client/src/lib.rs
@@ -1,9 +1,8 @@
 pub use base_client::{
     AuthIntegration, BaseClient, BaseClientBuilder, ClientBuildError, DEFAULT_CONNECT_TIMEOUT,
     DEFAULT_MAX_REDIRECTS, DEFAULT_READ_TIMEOUT, DEFAULT_READ_TIMEOUT_UPLOAD, DEFAULT_RETRIES,
-    ExtraMiddleware, RedirectClientWithMiddleware, RedirectPolicy, RequestBuilder, RetriableError,
-    RetryParsingError, RetryState, UvRetryableStrategy, fetch_with_url_fallback,
-    retryable_on_request_failure,
+    ExtraMiddleware, RedirectClientWithMiddleware, RedirectPolicy, RequestBuilder,
+    RetryParsingError, fetch_with_url_fallback,
 };
 pub use cached_client::{CacheControl, CachedClient, CachedClientError, DataWithCachePolicy};
 pub use error::{Error, ErrorKind, ProblemDetails, WrappedReqwestError};
@@ -13,6 +12,7 @@ pub use registry_client::{
     Connectivity, MetadataFormat, RegistryClient, RegistryClientBuilder, SimpleDetailMetadata,
     SimpleDetailMetadatum, SimpleIndexMetadata, VersionFiles,
 };
+pub use retry::{RetriableError, RetryState, UvRetryableStrategy, retryable_on_request_failure};
 pub use rkyvutil::{Deserializer, OwnedArchive, Serializer, Validator};
 
 mod base_client;
@@ -25,5 +25,6 @@ mod linehaul;
 mod middleware;
 mod registry_client;
 mod remote_metadata;
+mod retry;
 mod rkyvutil;
 mod tls;
